FEAT: userSeq로 사용자 정보 찾기 API

See merge request s09-webmobile1-sub2/S09P12B109!219

+++ server/rhythm4cuts/src/main/java/com/b109/rhythm4cuts/controller/MemberController.java
@@ -244,8 +244,8 @@ public class MemberController {
 
     // API. 유저 정보 조회 (userSeq)
     @GetMapping("/info/{userSeq}")
-    public ResponseEntity<?> getUserInfo(@PathVariable("userSeq") Long userSeq) throws SQLException {
-        UserDto userDto = userService.findById(userSeq);
+    public ResponseEntity<?> getUserInfo(@PathVariable("userSeq") int userSeq) throws SQLException {
+        UserDto userDto = userService.findById((long) userSeq);
 
         Map<String, Object> res = new HashMap<>();
